Refactored /api/teams/.
