--- Server/api/teams/handler/id/requestJoinToTeam.js
@@ -1,20 +1,0 @@
-'use strict';
-
-const Boom = require('boom');
-const Team = require('../../models/Team');
-const Code = require('../../../../config/errorCodes');
-
-module.exports = (req, res) => {
-  Team.findById(req.params.id, (err, team) => {
-    if (team.requests.length >= 10) {
-      res(Code.maxRequestsReached);
-    }else {
-      if (team.requests.indexof(req.payload.user) >= 0) {
-        res(Code.alreadyRequested);
-      }else {
-        team.requests.push({user: req.payload.user, role: req.payload.role, msg: req.payload.msg});
-        res({msg: 'Success'}).code(200);
-      }
-    }
-  });
-};
