--- Server/api/teams/handler/index.js
@@ -1,9 +1,0 @@
-module.exports = {
-  addTeam: require('./addTeam'),
-  getTeams: require('./getTeams'),
-  getTeam: require('./id/getTeam'),
-  updateTeam: require('./id/updateTeam'),
-  deleteTeam: require('./id/deleteTeam'),
-  addUserToTeam: require('./id/addUserToTeam'),
-  requestJoinToTeam: require('./id/requestJoinToTeam')
-};
