--- Server/bin/start
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-npm start
